+++ docs/code-examples/clear_simple.cpp
@@ -30,7 +30,7 @@ int main() {
         {200, 0, 200}};
 
     // Log a handful of arrows.
-    for (int i = 0; i < vectors.size(); ++i) {
+    for (size_t i = 0; i < vectors.size(); ++i) {
         auto entity_path = "arrows/" + std::to_string(i);
         rec.log(
             entity_path.c_str(),
@@ -39,7 +39,7 @@ int main() {
     }
 
     // Now clear them, one by one on each tick.
-    for (int i = 0; i < vectors.size(); ++i) {
+    for (size_t i = 0; i < vectors.size(); ++i) {
         auto entity_path = "arrows/" + std::to_string(i);
         rec.log(entity_path.c_str(), rr::Clear::FLAT);
     }
